Refacto to FSParser + Add support on FAT
