+++ tika-app/pom.xml
@@ -76,7 +76,7 @@
     </dependency>
         <dependency>
             <groupId>org.apache.tika</groupId>
-            <artifactId>tika-parser-ntfs-module</artifactId>
+            <artifactId>tika-parser-fs-module</artifactId>
             <version>${project.version}</version>
         </dependency>
