+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-modules/pom.xml
@@ -78,7 +78,7 @@
     <module>tika-parser-news-module</module>
     <module>tika-parser-crypto-module</module>
     <module>tika-parser-webarchive-module</module>
-    <module>tika-parser-ntfs-module</module>
+    <module>tika-parser-fs-module</module>
   </modules>
   <build>
     <pluginManagement>
