+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-modules/tika-parser-fs-module/src/main/resources/META-INF/services/org.apache.tika.detect.Detector
@@ -1,0 +1,2 @@
+org.apache.tika.parser.fs.NTFSDetector
+org.apache.tika.parser.fs.FATDetector
