+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-modules/tika-parser-fs-module/src/main/resources/META-INF/services/org.apache.tika.parser.Parser
@@ -1,0 +1,1 @@
+org.apache.tika.parser.fs.FSParser
