--- tika-parsers/tika-parsers-standard/tika-parsers-standard-modules/tika-parser-ntfs-module/src/main/resources/META-INF/services/org.apache.tika.parser.Parser
@@ -1,1 +1,0 @@
-org.apache.tika.parser.ntfs.NTFSParser
