+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-package/pom.xml
@@ -147,7 +147,7 @@
     </dependency>
     <dependency>
             <groupId>org.apache.tika</groupId>
-            <artifactId>tika-parser-ntfs-module</artifactId>
+            <artifactId>tika-parser-fs-module</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
